# web/openapi.yaml
basePath: /api/v1
definitions:
...
    profile.ProfilePaginatedResponse:
        properties:
            pagination:
                additionalProperties: true
                type: object
            profile:
                items:
                    $ref: '#/definitions/profile.ProfileResponse'
                type: array
        type: object
    profile.ProfileResponse:
        properties:
            category:
                $ref: '#/definitions/response.CategoryInfo'
            category_id:
                type: string
            content:
                type: string
            created_at:
                description: ISO 8601 format
                type: string
            id:
                type: string
            section_endpoint:
                type: string
            section_name:
                type: string
            state:
                type: boolean
            updated_at:
                description: ISO 8601 format
                type: string
        type: object
    profile.ProfileSectionsResponse:
        properties:
            section_names:
                items:
                    type: string
                type: array
        type: object
    response.CategoryInfo:
        properties:
            id:
                type: string
            name:
                type: string
        type: object
    response.ErrorResponse:
        properties:
            error:
                type: string
            success:
                type: boolean
        type: object
...
    response.MessageResponse:
        properties:
            message:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: REST API for the Desa (village) information system. Implements Clean Architecture with strict dependency rules, JWT (HS256) authentication, Casbin RBAC, PostgreSQL persistence, and rate limiting.
    title: Desa API
    version: "1.0"
paths:
...
    /profile:
        get:
            description: 'Filter by section_name, state, and q (search section_name + content). RBAC: profile:read.'
            parameters:
                - description: Section name
                  in: query
                  name: section_name
                  type: string
                - description: Enabled state
                  in: query
                  name: state
                  type: boolean
                - description: Search section_name or content
                  in: query
                  name: q
                  type: string
                - description: Page
                  in: query
                  name: page
                  type: integer
                - description: Page size
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfilePaginatedResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: List profile sections (admin)
            tags:
                - profile
        post:
            consumes:
                - application/json
            description: 'Required: content, section_name, section_endpoint (must start with /). Optional: state. RBAC: profile:write.'
            parameters:
                - description: '{content, section_name, section_endpoint, state}'
                  in: body
                  name: request
                  required: true
                  schema:
                    additionalProperties: true
                    type: object
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/profile.ProfileResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Create profile section (admin)
            tags:
                - profile
    /profile/{id}:
        delete:
            description: 'RBAC: profile:write.'
            parameters:
                - description: Profile UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/response.MessageResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Delete profile section (admin)
            tags:
                - profile
        get:
            description: 'RBAC: profile:read.'
            parameters:
                - description: Profile UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfileResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get profile section by ID (admin)
            tags:
                - profile
        put:
            consumes:
                - application/json
            description: 'RBAC: profile:write.'
            parameters:
                - description: Profile UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfileResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update profile section (admin)
            tags:
                - profile
...
    /profile/sections/names:
        get:
            description: 'RBAC: profile:read.'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfileSectionsResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: List distinct section names (admin)
            tags:
                - profile
...
    /public/profile/{id}:
        get:
            description: Public read.
            parameters:
                - description: Profile UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfileResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: Get profile section (public)
            tags:
                - profile
...
    /public/profile/list:
        get:
            description: Public read.
            parameters:
                - description: Page
                  in: query
                  name: page
                  type: integer
                - description: Page size
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfilePaginatedResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: List profile sections (public)
            tags:
                - profile
...
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'JWT bearer token. Format: `Bearer {token}`'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...